Создаст компонент Location
