+++ src/components/location/location-data.tsx
@@ -1,0 +1,8 @@
+export const cities: string[] = [
+  'Paris',
+  'Cologne',
+  'Brussels',
+  'Amsterdam',
+  'Hamburg',
+  'Dusseldorf',
+];
